Refactor component imports in layout and page files
