+++ src/app/[locale]/page.tsx
@@ -1,9 +1,11 @@
-import ChapterOne from '@/components/ChapterOne';
-import ChapterTwo from '@/components/ChapterTwo';
-import Landing from '@/components/Landing';
-import ChapterTwo2 from '@/components/ChapterTwo2';
-import ChapterThree from '@/components/ChapterThree';
-import Footer from '@/components/Footer';
+import {
+  ChapterOne,
+  ChapterTwo,
+  Landing,
+  ChapterThree,
+  ChapterTwo2,
+  Footer,
+} from '@/components/';
 import { useTranslations } from 'next-intl';
 import { unstable_setRequestLocale } from 'next-intl/server';
 import Intro from '@/components/Intro';
