+++ src/components/index.tsx
@@ -1,0 +1,15 @@
+export { default as ChapterOne } from './ChapterOne';
+export { default as ChapterThree } from './ChapterThree';
+export { default as ChapterTwo } from './ChapterTwo';
+export { default as ChapterTwo2 } from './ChapterTwo2';
+export { default as Footer } from './Footer';
+export { default as GameBoy } from './GameBoy';
+export { default as Icons } from './Icons';
+export { default as Intro } from './Intro';
+export { default as Landing } from './Landing';
+export { default as LocaleSwitcher } from './LocaleSwitcher';
+export { default as ModeToggle } from './ModeToggle';
+export { default as Navbar } from './Navbar';
+export { default as ThemeProvider } from './ThemeProvider';
+export { default as TitleWithDynamicBG } from './TitleWithDynamicBG';
+export { default as Video } from './Video';
